+++ src/components/Footer.jsx
@@ -1,0 +1,26 @@
+// Footer.jsx
+function Footer() {
+  return (
+    <>
+      <div className="contact">
+        <a href="https://discord.gg/xen4wNYMGt" className="contact-link">
+          Join the Discord
+        </a>
+        <p></p>
+        <a href="mailto:topcitysoftware@gmail.com" className="contact-link">
+          Contact Me
+        </a>
+        <p>BK Software LLC</p>
+      </div>
+
+      <div className="contributing">
+        <p>Interested in contributing? Check out the project on GitHub!</p>
+        <a href="https://github.com/BrettKulp/buzzbowl" target="_blank" className="contact-link">
+          GitHub Repository
+        </a>
+      </div>
+    </>
+  )
+}
+
+export default Footer